Node#GetLastCheckinTime should return a uint64, not a uint32

+++ lib/mrg/grid/config/Node.rb
@@ -74,7 +74,7 @@ module Mrg
         end
         
         expose :GetLastCheckinTime do |args|
-          args.declare :time, :uint32, :out, {}
+          args.declare :time, :uint64, :out, {}
         end
         
         # GetConfig
